Make it possible to exclude the admin account from internal metrics
